move menu into a seperate file
