+++ packages/react-components/src/components/ChatMessage/ChatMessageComponentAsMessageBubble.tsx
@@ -1,16 +1,11 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { IStyle, mergeStyles, Theme } from '@fluentui/react';
-import { Chat, Text, MoreIcon, MenuProps, Ref, ComponentSlotStyle } from '@fluentui/react-northstar';
+import { IStyle, mergeStyles } from '@fluentui/react';
+import { Chat, Text, ComponentSlotStyle } from '@fluentui/react-northstar';
 import { _formatString } from '@internal/acs-ui-common';
 import React, { useCallback, useRef, useState } from 'react';
-import {
-  chatActionsCSS,
-  iconWrapperStyle,
-  chatMessageEditedTagStyle,
-  chatMessageDateStyle
-} from '../styles/ChatMessageComponent.styles';
+import { chatMessageEditedTagStyle, chatMessageDateStyle } from '../styles/ChatMessageComponent.styles';
 import { formatTimeForChatMessage, formatTimestampForChatMessage } from '../utils/Datetime';
 import { useIdentifiers } from '../../identifiers/IdentifierProvider';
 import { useTheme } from '../../theming';
@@ -19,6 +14,7 @@ import { ChatMessageActionFlyout } from './ChatMessageActionsFlyout';
 import { ChatMessageContent } from './ChatMessageContent';
 import { ChatMessage } from '../../types/ChatMessage';
 import { MessageThreadStrings } from '../MessageThread';
+import { chatMessageActionMenuProps } from './ChatMessageActionMenu';
 
 type ChatMessageComponentAsMessageBubbleProps = {
   message: ChatMessage;
@@ -128,57 +124,4 @@ export const ChatMessageComponentAsMessageBubble = (props: ChatMessageComponentA
   );
 
   return chatMessage;
-};
-
-type ChatMessageActionMenuProps = MenuProps & {
-  showActionMenu?: boolean | undefined;
-};
-
-/**
- * Props for the Chat.Message action menu.
- * This is the 3 dots that appear when hovering over one of your own chat messages.
- */
-const chatMessageActionMenuProps = (menuProps: {
-  /** Whether the action menu button is enabled, if not this will always return undefined */
-  enabled: boolean;
-  /** Whether to force showing the action menu button - this has no effect if the action menu button is not enabled */
-  forceShow: boolean;
-  menuButtonRef: React.MutableRefObject<HTMLElement | null>;
-  onActionButtonClick: () => void;
-  theme: Theme;
-}): ChatMessageActionMenuProps | undefined => {
-  if (!menuProps.enabled) {
-    return undefined;
-  }
-
-  const menuClass = mergeStyles(chatActionsCSS, {
-    'ul&': { boxShadow: menuProps.theme.effects.elevation4, backgroundColor: menuProps.theme.palette.white }
-  });
-
-  const actionMenuProps: ChatMessageActionMenuProps = {
-    showActionMenu: menuProps.forceShow === true ? true : undefined,
-    iconOnly: true,
-    activeIndex: -1,
-    className: menuClass,
-    onItemClick: () => menuProps.onActionButtonClick(),
-    items: [
-      {
-        children: (
-          <Ref innerRef={menuProps.menuButtonRef}>
-            <MoreIcon
-              className={iconWrapperStyle}
-              {...{
-                outline: true
-              }}
-            />
-          </Ref>
-        ),
-
-        key: 'menuButton',
-        indicator: false
-      }
-    ]
-  };
-
-  return actionMenuProps;
 };
